feat(buildtool): added publish_halyard and publish_spinnaker
Removed the nebula decoy because nebula was publishing it
as the source to bintray. Instead I remove the conflicting tags
from the local repository as the original scripts had.

Refactored GitRemoteRepository into GitRepositorySpec which contains
the local_repo and origin URLs in addition to upstream URL (though upstream
isnt used by the buildtool). This was primarily done as a saftey mechanism
to make it difficult to use the default spinnaker origin repository by
accident. The repository spec can only be created dynamically, and uses the
configured github_user to derive the origin repository so authoritative repos
are unlikely to be hardcoded in. This also removes the need to pass git_dir
around or infer what it should be.

Added default $HOME/.spinnaker/buildtool.yml as base configuration file for
convienience and additional safety from using the wrong repo or bypassing
safety configurations by forgetting to pass an explicit configuration file.

Instrumentation is more pervasive, particularly around errors. All internal
exceptions result in counter increments, so all errors are captured by
monitoring. Where exceptions are thrown, they try to at least identify the
culprit if not the actual cause.

Remove Stackdriver and Prometheus support, added InfluxDB which supports
this tool's use case.

Refactored the build_command.py to break out halyard, debians, rpms, and
containers as separate modules to make it more readable.

Factored out gradle_support and hal_support modules to improve readability.
Factored out subprocess util functions from the general ones to avoid a
circular dependency with the util and error module when generating errors
in subprocesses.

Distinguish between specifying sources via BOMs (BomSourceCodeManager)
vs via git branches (BranchSourceCodeManager). RepositoryCommands are
defined using one or the other. This also means the BOM is generated first,
then the artifacts are built off that.

Some RepositoryCommands based on BOMs might also use a BranchSourceCodeManager
where they also operate on repositories outside the BOM. In particular when
publishing documentation for a release defined by a BOM, the documentation
is published to a documentation repository outside the BOM.